Migrate to ATOMS tests
