+++ spec/model/arubainstant_spec.rb
@@ -1,4 +1,5 @@
 require_relative 'model_helper'
+require_relative 'atoms'
 
 describe 'model/ArubaInstant' do
   before(:each) do
@@ -8,24 +9,9 @@ describe 'model/ArubaInstant' do
                                model: 'arubainstant')
   end
 
-  it "matches different prompts" do
-    # Virtual controller - ArubaOS (MODEL: 515), Version 8.10.0.7 LSR
-    _('AAAA-AP123456# ').must_match ArubaInstant.prompt
-  end
-
-  it 'runs on IAP515 with 8.10.0.6' do
-    mockmodel = MockSsh.new('examples/device-simulation/yaml/arubainstant_IAP515_8.10.0.6_VWLC.yaml')
-    Net::SSH.stubs(:start).returns mockmodel
-
-    status, result = @node.run
-
-    _(status).must_equal :success
-    _(result.to_cfg).must_equal mockmodel.oxidized_output
-  end
-
   it 'removes secrets' do
     Oxidized.config.vars.remove_secret = true
-    mockmodel = MockSsh.new('examples/device-simulation/yaml/arubainstant_IAP515_8.10.0.6_VWLC.yaml')
+    mockmodel = MockSsh.new(ATOMS::TestOutput.new('arubainstant', 'IAP515_8.10.0.6_VWLC'))
     Net::SSH.stubs(:start).returns mockmodel
 
     status, result = @node.run
