+++ spec/model/data/aoscx:generic:prompt.yaml
@@ -1,0 +1,3 @@
+---
+pass:
+  - 'LAB-SW1234# '
