+++ spec/model/data/arubainstant:generic:prompt.yaml
@@ -1,0 +1,2 @@
+pass:
+  - 'AAAA-AP123456# '
