+++ spec/model/data/asa:generic:prompt.yaml
@@ -1,0 +1,3 @@
+pass:
+  - "\rLAB-ASA12-Oxidized-IPv6>"
+  - "\rLAB-ASA12-Oxidized-IPv6#"
