+++ spec/model/data/garderos:generic:prompt.yaml
@@ -1,0 +1,2 @@
+pass:
+  - "LAB-R1234_Garderos# "
