+++ spec/model/data/ios:C9200L-24P-4G_17.09.04a:output.txt
