+++ spec/model/data/ios:C9800-L-F-K9_17.06.05:output.txt
