--- spec/model/data/ios:iosxe_C9200L-24P-4G_17.09.04a:output.txt
@@ -1,204 +1,0 @@
-! Cisco IOS XE Software, Version 17.09.04a
-!\x20
-! Image: Software: CAT9K_LITE_IOSXE, 17.9.4a, RELEASE SOFTWARE (fc3)
-! Image: Compiled: Fri 20-Oct-23 10:33 by mcpre
-! Image: flash:cat9k_lite_iosxe.17.09.04a.SPA.bin
-! Chassis type: C9200L-24P-4G
-! Memory: main 519464K/3071K
-! Processor ID: JAE24FFFFFF
-! CPU: ARM64
-! Memory: nvram 2048K
-!\x20
-! VTP: VTP Version capable             : 1 to 3
-! VTP: VTP version running             : 1
-! VTP: VTP Domain Name                 : Oxidized
-! VTP: VTP Pruning Mode                : Disabled (Operationally Disabled)
-! VTP: VTP Traps Generation            : Disabled
-! VTP: Device ID                       : 40f0.7800.0000
-! VTP: Feature VLAN:
-! VTP: --------------
-! VTP: VTP Operating Mode                : Transparent
-! VTP: Maximum VLANs supported locally   : 1005
-! VTP: Number of existing VLANs          : 10
-! VTP: Configuration Revision            : 0
-! VTP: MD5 digest                        : 0x35 0x00 0x00 0x00 0x00 0x00 0x7F 0xB4\x20
-! VTP:                                     0x07 0x00 0x00 0x00 0x00 0x00 0x09 0x6D\x20
-!\x20
-! NAME: \"c92xxL Stack\", DESCR: \"c92xxL Stack\"
-! PID: C9200L-24P-4G     , VID: V01  , SN: JAE24FFFFFF
-!\x20
-! NAME: \"Switch 1\", DESCR: \"C9200L-24P-4G\"
-! PID: C9200L-24P-4G     , VID: V01  , SN: JAE24FFFFFF
-!\x20
-! NAME: \"Switch 1 - Power Supply A\", DESCR: \"Switch 1 - Power Supply A\"
-! PID: PWR-C5-600WAC     , VID: V02  , SN: QCS24XXXXXX
-!\x20
-!\x20
-!
-! Last configuration change at 09:37:40 CEST Tue Aug 20 2024 by oxidized
-! NVRAM config last updated at 09:37:37 CEST Tue Aug 20 2024 by oxidized
-!
-version 17.9
-service timestamps debug datetime localtime show-timezone year
-service timestamps log datetime localtime show-timezone year
-service password-encryption
-!
-hostname LAB-SW123_9200L
-!
-!
-vrf definition Mgmt-vrf
- !
- address-family ipv4
- exit-address-family
- !
- address-family ipv6
- exit-address-family
-!
-logging buffered 65536 informational
-no logging console
-aaa new-model
-!
-!
-aaa authentication login default group tacacs+ local
-aaa authentication enable default group tacacs+ enable
-aaa authorization exec default group tacacs+ if-authenticated\x20
-! ...
-no aaa accounting system guarantee-first
-!
-!
-aaa session-id common
-!
-!
-!
-clock timezone CET 1 0
-clock summer-time CEST recurring last Sun Mar 2:00 last Sun Oct 2:00
-boot system switch all flash:cat9k_lite_iosxe.17.09.05.SPA.bin
-switch 1 provision c9200l-24p-4g
-!
-!
-!
-!
-!
-ip name-server 10.42.0.1 10.42.0.2
-no ip domain lookup
-ip domain name oxidized.local
-!
-!
-!
-login on-success log
-vtp domain Oxidized
-vtp mode transparent
-vtp version 1
-!
-!
-spanning-tree mode rapid-pvst
-spanning-tree logging
-spanning-tree portfast bpduguard default
-spanning-tree extend system-id
-!
-!
-errdisable recovery cause udld
-errdisable recovery cause bpduguard
-errdisable recovery cause security-violation
-errdisable recovery cause dhcp-rate-limit
-errdisable recovery cause arp-inspection
-errdisable recovery cause loopback
-!
-enable secret 9 $9$xxxSECRET_ENABLExxxxxxxxxxxxxx
-!
-username oxidized secret 9 $9$xxxxxxxxxxSECRET_OXIDIZEDxxxxxxxxx
-username rocks secret 9 $9$xxxxxxxxxxSECRET_OXIDIZEDxxxxxxxxx
-!
-transceiver type all
- monitoring
-!
-vlan 2
- name two
-!
-vlan 3
- name three
-!
-!
-interface GigabitEthernet0/0
- vrf forwarding Mgmt-vrf
- no ip address
-!
-interface GigabitEthernet1/0/1
- description oxidized test one
- switchport access vlan 2
- switchport mode access
- no snmp trap link-status
- spanning-tree portfast
-!
-interface GigabitEthernet1/0/2
- description oxidized two
- switchport access vlan 3
- switchport mode access
-!
-! ...
-!
-interface GigabitEthernet1/0/24
- description oxodized trunk
- switchport trunk allowed vlan 2,3
- switchport mode trunk
- switchport nonegotiate
-!
-interface GigabitEthernet1/1/1
-!
-interface GigabitEthernet1/1/2
-!
-interface GigabitEthernet1/1/3
-!
-interface GigabitEthernet1/1/4
-!
-interface Vlan1
- no ip address
- shutdown
-!
-interface Vlan2
- description network management
- ip address 10.42.11.42 255.255.255.0
-!
-ip default-gateway 10.41.11.1
-no ip http server
-no ip http secure-server
-ip ftp source-interface Vlan2
-ip tftp source-interface Vlan2
-ip tacacs source-interface Vlan2\x20
-ip ssh version 2
-!
-!
-ip access-list standard 5
- 10 permit 10.0.0.0 0.0.0.255
- 20 deny   any
-!
-snmp-server community public RW 5
-snmp-server location Here and there
-snmp-server host 10.42.0.33 version 2c AAAAAAAAAABBBBBBBBBB\x20
-snmp-server host 10.42.0.34 version 2c AAAAAAAAAABBBBBBBBBB\x20
-snmp-server host 10.42.0.35 vrf Mgmt-vrf informs version 2c AAAAAAAAAABBBBBBBBBB  tty vtp
-snmp-server host 10.42.0.36 vrf Mgmt-vrf informs version 3 auth oxidized  bfd bridge transceiver
-!
-snmp mib expression owner nat name 1
-  expression <value removed>
-snmp mib expression owner oxidized name rocks
-  expression <value removed>
-!
-!
-!
-banner exec ^C
-### EXAMPLE IOS-XE 17.09.04a config on a C9200L-24P-4G ###
-Retrieved 2024-09-09 by @robertcheramy
-### ###
-^C
-banner login ^C
-.
-Login banner.
-.
-^C
-!
-!
-!
-!
-!
-end\n
