+++ spec/model/data/linuxgeneric:generic:prompt.yaml
@@ -1,0 +1,2 @@
+pass:
+  - 'robert@gap:~$ '
