--- spec/model/data/riverbed:915:output.txt
@@ -1,39 +1,0 @@
-! Date of version: YYYY-MM-DD HH:MM:SS UTC
-! Serial: SERIAL1234567
-! Product name: rbt_sh
-! Product release: X.Y.Z
-! Build ID: #1234
-! Build date: YYYY-MM-DD HH:MM:SS
-! Build arch: x86_64
-! Built by: builduser@buildhost
-! Product model: ModelName (ModelVariant)
-! Number of CPUs: XX
-! Hardware revision: RevX
-! Mainboard: MainboardDescription
-! Slot 0: ModuleDescription0
-! Slot 1: ModuleDescription1
-! Slot 2: ModuleDescription2
-! System led: Color
-##
-## Network interface configuration
-##
-interface inpath0_0 description "Interface Description"
-no interface inpath0_0 dhcp
-no interface inpath0_0 dhcp dynamic-dns
-no interface inpath0_0 dhcpv6
-no interface inpath0_0 dhcpv6 dynamic-dns
-##
-## Secure device access configuration
-##
-access enable
-##
-## Port Labels
-##
-port-label "Label" port "PortNumbers"
-tacacs-server host 192.0.2.254 key <secret hidden>
-username admin password <secret hidden>
-snmp-server community <configuration removed>
-ntp server 192.0.2.100 key <secret hidden>
-ip security shared secret <secret hidden>
-!
-end
