--- spec/model/data/riverbed:915:simulation.yaml
@@ -1,83 +1,0 @@
----
-init_prompt: |-
-  Riverbed SteelHead
-  Last login: Fri Nov 1 05:55:01 2024 from 192.0.2.1
-  steelhead >
-commands:
-  enable: |-
-    enable
-    steelhead #
-  terminal length 0: |-
-    terminal length 0
-    steelhead #
-  terminal width 1024: |-
-    terminal width 1024
-    steelhead #
-  show version: |-
-    show version
-    Product name:      rbt_sh
-    Product release:   X.Y.Z
-    Build ID:          #1234
-    Build date:        YYYY-MM-DD HH:MM:SS
-    Build arch:        x86_64
-    Built by:          builduser@buildhost
-
-    Uptime:            100d 10h 10m 10s
-
-    Product model:     ModelName (ModelVariant)
-    System memory:     XXXX MB used / XXXX MB free / XXXX MB total
-    Number of CPUs:    XX
-    CPU load averages: X.XX / X.XX / X.XX
-    steelhead #
-  show hardware all: |-
-    show hardware all
-    Hardware revision: RevX
-    Mainboard:  MainboardDescription
-    Slot 0: ..........  ModuleDescription0
-    Slot 1: ..........  ModuleDescription1
-    Slot 2: ..........  ModuleDescription2
-    System led: Color
-    steelhead #
-  show info: |-
-    show info
-    Current User:      admin
-
-    Status:            Normal
-    Config:            config.bak
-    Appliance Up Time: 100d 10h 10m 10s
-    Service Up Time:   90d 20h 20m 20s
-    Managed by CMC:    no
-    Temperature (C):   XX
-
-    Serial:            SERIAL1234567
-    Model:             ModelName (ModelVariant)
-    Revision:          RevX
-    Version:           X.Y.Z
-    steelhead #
-  show running-config: |-
-    show running-config
-    ##
-    ## Network interface configuration
-    ##
-    interface inpath0_0 description "Interface Description"
-    no interface inpath0_0 dhcp
-    no interface inpath0_0 dhcp dynamic-dns
-    no interface inpath0_0 dhcpv6
-    no interface inpath0_0 dhcpv6 dynamic-dns
-    ##
-    ## Secure device access configuration
-    ##
-    access enable
-    ##
-    ## Port Labels
-    ##
-    port-label "Label" port "PortNumbers"
-    tacacs-server host 192.0.2.254 key secretkey
-    username admin password 7 secretpassword
-    snmp-server community public
-    ntp server 192.0.2.100 key 1
-    ip security shared secret mysecret
-    !
-    end
-    steelhead #
-  exit: ""
