+++ spec/model/ios_spec.rb
@@ -1,4 +1,5 @@
 require_relative 'model_helper'
+require_relative 'atoms'
 
 describe 'model/IOS' do
   before(:each) do
@@ -8,44 +9,9 @@ describe 'model/IOS' do
                                model: 'ios')
   end
 
-  it 'matches different prompts' do
-    _('LAB-SW123_9200L#').must_match IOS.prompt
-    _('OXIDIZED-WLC1#').must_match IOS.prompt
-  end
-
-  it 'runs on C9200L-24P-4G with IOS-XE 17.09.04a' do
-    mockmodel = MockSsh.new('examples/device-simulation/yaml/iosxe_C9200L-24P-4G_17.09.04a.yaml')
-    Net::SSH.stubs(:start).returns mockmodel
-
-    status, result = @node.run
-
-    _(status).must_equal :success
-    _(result.to_cfg).must_equal mockmodel.oxidized_output
-  end
-
-  it 'runs on C9800-L-F-K9 with IOS-XE 17.06.05' do
-    mockmodel = MockSsh.new('examples/device-simulation/yaml/iosxe_C9800-L-F-K9_17.06.05.yaml')
-    Net::SSH.stubs(:start).returns mockmodel
-
-    status, result = @node.run
-
-    _(status).must_equal :success
-    _(result.to_cfg).must_equal mockmodel.oxidized_output
-  end
-
-  it 'runs on ASR920 with IOS 16.8.1b' do
-    mockmodel = MockSsh.new('examples/device-simulation/yaml/asr920_16.8.1b.yaml')
-    Net::SSH.stubs(:start).returns mockmodel
-
-    status, result = @node.run
-
-    _(status).must_equal :success
-    _(result.to_cfg).must_equal mockmodel.oxidized_output
-  end
-
   it 'removes secrets' do
     Oxidized.config.vars.remove_secret = true
-    mockmodel = MockSsh.new('examples/device-simulation/yaml/iosxe_C9200L-24P-4G_17.09.04a.yaml')
+    mockmodel = MockSsh.new(ATOMS::TestOutput.new('ios', 'C9200L-24P-4G_17.09.04a'))
     Net::SSH.stubs(:start).returns mockmodel
 
     status, result = @node.run
@@ -58,7 +24,7 @@ describe 'model/IOS' do
 
   it 'removes secrets from IOS-XE WLCs' do
     Oxidized.config.vars.remove_secret = true
-    mockmodel = MockSsh.new('examples/device-simulation/yaml/iosxe_C9800-L-F-K9_17.06.05.yaml')
+    mockmodel = MockSsh.new(ATOMS::TestOutput.new('ios', 'C9800-L-F-K9_17.06.05'))
     Net::SSH.stubs(:start).returns mockmodel
 
     status, result = @node.run
